Changed return type of main() from LONG to int
git-svn-id: https://svn.aros.org/svn/aros/trunk@6661 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/c/AddDatatypes.c
@@ -156,11 +156,11 @@ struct StackVars
 *
 */
 
-LONG main(void)
+int main(void)
 {
     struct StackVars  vars;
     struct StackVars *sv;
-    ULONG             result = RETURN_FAIL;
+    int    result = RETURN_FAIL;
 	
     memset(&vars, 0, sizeof(struct StackVars));
     sv = &vars;
